fix: linting issues
